修复：
- 解决目录大小写问题
- 移除遗留代码
- 清理eventmanager
