+++ src/Application.Core/Scripting/Events/AbstractEventManager.cs
@@ -1,13 +1,10 @@
 using Application.Core.Channel;
-using Application.Core.Game.Life;
 using Application.Core.Game.Maps;
 using Application.Core.model;
-using Application.Core.scripting.Events;
 using Application.Core.scripting.Events.Abstraction;
 using Application.Core.scripting.Events.Instances;
 using Application.Core.scripting.Events.Templates;
 using Application.Resources.Messages;
-using Application.Shared.Events;
 using Application.Utility.Performance;
 using Application.Utility.Tickables;
 using server.quest;
@@ -271,7 +268,7 @@ namespace Application.Core.Scripting.Events
             instantiateQueuedInstance();
         }
 
-        public virtual string? HandleCreateInstanceResult(CreateInstanceResult r, IChannelClient c)
+        public string? HandleCreateInstanceResult(CreateInstanceResult r, IChannelClient c)
         {
             return Template.HandleCreateInstanceResult(r, c);
         }
@@ -322,244 +319,14 @@ namespace Application.Core.Scripting.Events
             var eim = newInstance(Name + lobbyId);
             eim.setProperty("level", level);
 
-            OnSetup(eim, level, lobbyId);
-            respawnStages(eim);
+            Template.OnSetup(eim, level, lobbyId);
+            Template.respawnStages(eim);
             eim.startEventTimer(EventTime * 1000);
-            setEventRewards(eim);
+            Template.setEventRewards(eim);
 
             return eim;
         }
 
-        protected virtual void OnSetup(AbstractEventInstanceManager eim, int level, int lobbyId) { }
-
-        public virtual void AfterSeup(AbstractEventInstanceManager eim) { }
-
-        protected virtual void respawnStages(AbstractEventInstanceManager eim) { }
-
-        protected virtual void setEventRewards(AbstractEventInstanceManager eim) { }
-
-        public virtual List<Player> GetEligibleParty(Player leader)
-        {
-            var members = leader.getPartyMembersOnSameMap();
-
-            if (members.Count >= MinCount
-                && members.Count <= MaxCount
-                && members.All(x => x.Level >= MinLevel && x.Level <= MaxLevel))
-            {
-                return members;
-            }
-            return [];
-        }
-
-        public string GetRequirementDescription(IChannelClient client)
-        {
-            var countRange = MinCount == MaxCount ? MinCount.ToString() : MinCount + " ~ " + MaxCount;
-            var levelRange = MinLevel == MaxLevel ? MinLevel.ToString() : MinLevel + " ~ " + MaxLevel;
-            return client.CurrentCulture.GetScriptTalkByKey(nameof(ScriptTalk.PartyQuest_Requirement),
-                countRange,
-                levelRange,
-                (EventTime / 60).ToString());
-        }
-
-        protected virtual void End(AbstractEventInstanceManager eim)
-        {
-            eim.Dispose();
-        }
-
-        public virtual void OnTimeOut(AbstractEventInstanceManager eim)
-        {
-            Template.OnTimeOut(eim);
-        }
-
-        public virtual void OnPlayerRegister(AbstractEventInstanceManager eim, Player chr)
-        {
-            Template.OnPlayerRegister(eim, chr);
-        }
-
-        public virtual void OnPlayerEntry(AbstractEventInstanceManager eim, Player chr)
-        {
-            Template.OnPlayerEntry(eim, chr);
-        }
-
-        public virtual void OnPlayerExit(AbstractEventInstanceManager eim, Player player)
-        {
-            Template.OnPlayerExit(eim, player);
-        }
-
-        public virtual void OnPlayerUnregister(AbstractEventInstanceManager eim, Player chr)
-        {
-            Template.OnPlayerUnregister(eim, chr);
-        }
-
-        public virtual void OnPlayerMapChanging(AbstractEventInstanceManager eim, Player player, int mapid)
-        {
-            Template.OnPlayerMapChanging(eim, player, mapid);
-        }
-
-        public virtual void OnPlayerMapChanged(AbstractEventInstanceManager eim, Player player, int mapid)
-        {
-            Template.OnPlayerMapChanged(eim, player, mapid);
-        }
-
-        public virtual void OnPlayerDied(AbstractEventInstanceManager eim, Player chr)
-        {
-            Template.OnPlayerDied(eim, chr);
-        }
-
-        public virtual void OnPlayerDisconnected(AbstractEventInstanceManager eim, Player player)
-        {
-            Template.OnPlayerDisconnected(eim, player);
-        }
-
-        public virtual void OnLeaderChanged(AbstractEventInstanceManager eim, Player leader)
-        {
-            Template.OnLeaderChanged(eim, leader);
-        }
-
-        public virtual void OnPlayerLeftParty(AbstractEventInstanceManager eim, Player player)
-        {
-            Template.OnPlayerLeftParty(eim, player);
-        }
-
-        public virtual void OnPartyDisband(AbstractEventInstanceManager eim)
-        {
-            Template.OnPartyDisband(eim);
-        }
-
-        public virtual bool OnPlayerRevive(AbstractEventInstanceManager eim, Player player)
-        {
-            return Template.OnPlayerRevive(eim, player);
-        }
-
-        public virtual bool IsEventTeamLackingNow(AbstractEventInstanceManager eim, bool leavingEventMap, Player quitter)
-        {
-            return Template.IsEventTeamLackingNow(eim, leavingEventMap, quitter);
-        }
-
-        public virtual void OnMobKilled(AbstractEventInstanceManager eim, Monster mob, ICombatantObject? killer)
-        {
-            Template.OnMobKilled(eim, mob, killer);
-        }
-
-        public virtual void OnMobRevive(AbstractEventInstanceManager eim, Monster mob)
-        {
-            Template.OnMobRevive(eim, mob);
-        }
-
-        public virtual void OnMobClear(AbstractEventInstanceManager eim, IMap map)
-        {
-            Template.OnMobClear(eim, map);
-        }
-
-        public virtual void OnFriendlyMobDamaged(AbstractEventInstanceManager eim, Monster mob, ICombatantObject? attacker, int damage)
-        {
-            Template.OnFriendlyMobDamaged(eim, mob, attacker, damage);
-        }
-
-        public virtual void OnFriendlyMobKilled(AbstractEventInstanceManager eim, Monster mob, ICombatantObject? killer)
-        {
-            Template.OnFriendlyMobKilled(eim, mob, killer);
-        }
-
-        public virtual void OnFriendlyMobDrop(AbstractEventInstanceManager eim, Monster mob)
-        {
-            Template.OnFriendlyMobDrop(eim, mob);
-        }
-
-        public virtual void ClearPQ(AbstractEventInstanceManager eim)
-        {
-            Template.ClearPQ(eim);
-        }
-
-        public virtual RewardOptions GetAllClearRewardOptions(Player chr, int point)
-        {
-            return new RewardOptions();
-        }
-
-        public virtual ClaimRewardResult GiveClearReward(AbstractEventInstanceManager eim, Player player, int point)
-        {
-            if (!eim.CanGiveReward(player, -1))
-            {
-                return ClaimRewardResult.Claimed;
-            }
-
-            var pool = AllClearRewards.GetValueOrDefault(eim.Level);
-            if (pool == null)
-            {
-                pool = AllClearRewards.Values.FirstOrDefault();
-                if (pool == null)
-                {
-                    return ClaimRewardResult.Success;
-                }
-            }
-
-            if (pool.ItemPool.Length > 0)
-            {
-                if (!eim.hasRewardSlot(player, eim.Level))
-                {
-                    return ClaimRewardResult.BagFull;
-                }
-
-                var item = Randomizer.Select(pool.ItemPool);
-                player.GainItem(item.ItemId, (short)item.Quantity, show: GainItemShow.ShowInChat);
-            }
-
-            var option = GetAllClearRewardOptions(player, point);
-            if (pool.ExpPool.Length > 0)
-            {
-                var baseExp = option.ExpPoolIndex == -1 ? Randomizer.Select(pool.ExpPool) : pool.ExpPool[option.ExpPoolIndex];
-                player.gainExp((int)(baseExp * option.FinalExpRate));
-            }
-
-            if (pool.MesoPool.Length > 0)
-            {
-                var baseMeso = option.MesoPoolIndex == -1 ? Randomizer.Select(pool.MesoPool) : pool.MesoPool[option.MesoPoolIndex];
-                player.GainMeso((int)(baseMeso * option.FinalMesoRate), GainItemShow.ShowInChat);
-            }
-
-            eim.SetRewardClaimed(player, -1);
-            return ClaimRewardResult.Success;
-        }
-
-        public virtual RewardOptions GetStageClearRewardOptions(Player chr, int stageMap)
-        {
-            return new RewardOptions();
-        }
-
-        public virtual ClaimRewardResult GiveStageClearReward(AbstractEventInstanceManager eim, Player player, int stageMap)
-        {
-            if (StageClearRewards.TryGetValue(stageMap, out var data))
-            {
-                var expExtraBonus = eim.Type == EventInstanceType.PartyQuest ? YamlConfig.config.server.PQ_BONUS_EXP_RATE : 1;
-                var option = GetStageClearRewardOptions(player, stageMap);
-                if (eim.CanGiveReward(player, stageMap))
-                {
-                    eim.SetRewardClaimed(player, stageMap);
-                    player.gainExp((int)(data.Exp * option.FinalExpRate), true, true);
-                    player.GainMeso((int)(data.Meso * option.FinalMesoRate), GainItemShow.ShowInChat);
-                    return ClaimRewardResult.Success;
-                }
-                return ClaimRewardResult.Claimed;
-            }
-            return ClaimRewardResult.Success;
-        }
-
-        public virtual void GiveStageClearRewardAll(AbstractEventInstanceManager eim, int stageMap)
-        {
-            if (StageClearRewards.TryGetValue(stageMap, out var data))
-            {
-                foreach (var player in eim.getPlayers())
-                {
-                    var option = GetStageClearRewardOptions(player, stageMap);
-                    if (eim.CanGiveReward(player, stageMap))
-                    {
-                        eim.SetRewardClaimed(player, stageMap);
-                        player.gainExp((int)(data.Exp * option.FinalExpRate), true, true);
-                        player.GainMeso((int)(data.Meso * option.FinalMesoRate), GainItemShow.ShowInChat);
-                    }
-                }
-            }
-        }
 
         class DelayedDisposeRequest : DelayedTickable
         {
