+++ src/Application.Core/Scripting/Events/AriantEventManager.cs
