+++ src/Application.Core/Scripting/Events/EventType.cs
