+++ src/Application.Core/Scripting/Events/Templates/AbstractGuildQuestTemplate.cs
