--- src/Application.Plugin.Events/Application.Plugin.Events.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <ProduceReferenceAssembly>False</ProduceReferenceAssembly>
-    <BaseOutputPath>..\Application.Host\bin\</BaseOutputPath>
-  </PropertyGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Application.Core\Application.Core.csproj" />
-	</ItemGroup>
-
-</Project>
