+++ src/Application.Plugin.Script/Npc/Balrog.cs
@@ -1,7 +1,7 @@
 using Application.Core.scripting.Events.Instances;
 using Application.Core.Scripting.Events;
 
-namespace Application.Plugin.Script.Events
+namespace Application.Plugin.Script.Npc
 {
     internal partial class NpcScript
     {
@@ -22,7 +22,7 @@ namespace Application.Plugin.Script.Events
             var expedition = em.GetOnlyEventInstanceManager<ExpeditionEventInstanceManager>();
             if (expedition == null)
             {
-                var selection = await AskMenu("#e#b<远征：" + expedBoss + ">\r\n#k#n" + em.GetRequirementDescription(c) + "\r\n\r\n你想组建一个远征队来挑战 #r" + expedBoss + "#k 吗？\r\n#b#L1#让我们开始吧！#l\r\n#L2#不，我想再等一会儿...#l");
+                var selection = await AskMenu("#e#b<远征：" + expedBoss + ">\r\n#k#n" + em.Template.GetRequirementDescription(c) + "\r\n\r\n你想组建一个远征队来挑战 #r" + expedBoss + "#k 吗？\r\n#b#L1#让我们开始吧！#l\r\n#L2#不，我想再等一会儿...#l");
 
                 if (selection == 1)
                 {
@@ -104,7 +104,7 @@ namespace Application.Plugin.Script.Events
                 }
                 else
                 {
-                    await SayOK(em.HandleJoinInstanceResult(em.JoinMember(expedition, getPlayer()), c));
+                    await SayOK(em.GetTemplate.HandleJoinInstanceResult(em.GetTemplate.JoinMember(expedition, getPlayer()), c));
                 }
             }
             else if (expedition.isInProgress())
