+++ src/Application.Plugin.Script/Npc/PQ_Ludi.cs
@@ -1,8 +1,9 @@
 using Application.Core.scripting.Events.Abstraction;
 using Application.Core.scripting.Events.Instances;
 using Application.Core.Scripting.Events;
+using Application.Plugin.Script.Events;
 
-namespace Application.Plugin.Script.Events
+namespace Application.Plugin.Script.Npc
 {
     internal partial class NpcScript
     {
@@ -11,7 +12,7 @@ namespace Application.Plugin.Script.Events
         {
             var em = GetEventManager(nameof(PQ_Ludi));
 
-            var option = await AskMenu($"#e#b<组队任务：时空裂缝>\r\n#k#n{em.GetRequirementDescription(c)}\r\n\r\n由于上方有极其危险的生物，你无法再往上走。你想要和队友合作完成任务吗？如果是，请让你的#b队长#k和我交谈。",
+            var option = await AskMenu($"#e#b<组队任务：时空裂缝>\r\n#k#n{em.Template.GetRequirementDescription(c)}\r\n\r\n由于上方有极其危险的生物，你无法再往上走。你想要和队友合作完成任务吗？如果是，请让你的#b队长#k和我交谈。",
                 ["我想参加组队任务。", "我想了解更多详情。"]
             );
             switch (option)
