+++ src/Application.Plugin.Script/Npc/Wedding.cs
@@ -1,6 +1,6 @@
 using Application.Shared.Constants.Map;
 
-namespace Application.Plugin.Script.Events
+namespace Application.Plugin.Script.Npc
 {
     internal partial class NpcScript
     {
